+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### vNEXT
 - Added example and demo for max height in dialog
+- Added expandable search component and options container
 
 ### v0.4.0
 - Introducing first pass of text-list-item component